requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.25.0
plotly>=5.18.0
yfinance>=0.2.36
alpaca-py>=0.20.0
lxml>=5.0.0
html5lib>=1.1
